api/requirements.txt
beautifulsoup4
tqdm
faiss-cpu
sentence-transformers
fastapi
uvicorn
pydantic
requests
python-multipart
numpy
